+++ src/main/java/de/samply/generator/TestDataGeneratorException.java
@@ -1,0 +1,7 @@
+package de.samply.generator;
+
+public class TestDataGeneratorException extends Exception{
+    public TestDataGeneratorException(Throwable cause) {
+        super(cause);
+    }
+}